Ehhhhhhh light changes are now properly cached :sweat_smile:
Oops I compared the crate::Color for caching, i.e. I compared the float
colors, not the quantized color to find out whether a certain light
needs to be transmitted or not.
Because I compared floats, and the color is ever changing in my pulse,
I was essentially not caching at all and dumping out 4800 MIDI messages
per second, per device

:sob: I'm sorry my dear Launchpads, I swear it won't happen again