S: use rapid update if lots of changes
This involves changing the DeviceSpec trait so that it accepts a
DeviceCanvas reference instead of just the Self::Output reference.
That way, implementors have access to the entire grid - even unchanged
pixels, which is important for the rapid update feature, in which I need
to update all LEDs of the grid and can't update _just_ the changed
pixels.