:pray::pray::pray: Canvas restructure
instead of the _new methods being the default ones and _old having the
prefix, now it's the other way around. the more common use case - which
are the *_old methods - now can be accessed without any prefix for
convenience, and the *_new methods get the prefix.

Well that was ONE thing, but I ALSO restructured the Canvas trait to
work on references instead of owned values!!! Big step that I'd have
thought I'd never take, a few days ago. But yeah here we are, and this
_allows implementing Index and IndexMut!!!!!!!!_ pure bliss