refactor: change Button80 into LogicalButton/PhysicalButton
Button80 used to contain the logic for translating between button
identifiers and button coordinates. In the implementation of the
Mini Mk3 it looks like the coordinate translation depends on
the type of LaunchPad, so its calculation should involve
the LaunchPad code.

In this PR, rename Button80 to LogicalButton (the button
identifiers for use with the low-level MIDI interface) and
split off PhysicalButton (the X, Y coordinates for use with
the Canvas). Translation between them happens via the `DeviceSpec`.

Since up until now the logic could be shared between all LaunchPad
implementations, add default implementations for the conversion
in `shared/mod.ts` and forward all current implementations there.
The implementation of the Mk3 will have its own additional logic.